refactor: rename connectivity/directed.py to general.py and clean up tests
Per Leif's review:
- Rename create/mesh/connectivity/directed.py -> general.py so the module
  sits at .mesh.connectivity.general alongside .flat and .hierarchical.
  'general' better describes the role (a general-purpose building block
  consumed by both flat and hierarchical) rather than the return type.
- Update all imports in __init__.py, flat.py, hierarchical.py and coords.py
  to reference .general instead of .directed
- Remove # noqa: F401 from coords.py import (it is a real internal import,
  not a backward-compat re-export)
- Remove TestReturnTypeAnnotations test class (static type checkers handle
  this, not runtime tests)
- Remove TestDirectedMeshGraphModuleLocation test class (was testing
  re-export wiring that no longer exists)
