+++ src/weather_model_graphs/create/mesh/connectivity/__init__.py
@@ -1,3 +1,3 @@
-from .directed import create_directed_mesh_graph
+from .general import create_directed_mesh_graph
 
 __all__ = ["create_directed_mesh_graph"]
